third_party/blink/web_tests/virtual/legacy-client-hints-no-fp-delegation/external/wpt/client-hints/accept-ch-change.https-expected.txt

This is a testharness.js-based test.
FAIL Accept-CH changes based on header assert_equals: expected "PASS" but got "FAIL"
Harness: the test ran to completion.
